+++ yeeun/animation/src/App.css
@@ -1,25 +1,38 @@
-body {
-  margin: 0;
-  padding: 0;
-  background-color: black;
-  height: 100vh;
+.App {
+  text-align: center;
+}
+
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
   display: flex;
-  justify-content: center;
+  flex-direction: column;
   align-items: center;
-}
-
-.logo-container {
-  display: flex;
   justify-content: center;
-  align-items: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
 }
 
-.netflix-logo {
-  display: flex;
+.App-link {
+  color: #61dafb;
 }
 
-.netflix-letter {
-  font-size: 6rem;
-  font-weight: bold;
-  color: red;
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
 }
